MIgrate to Nav3, Add tags infrastructure + homepage theming tweaks

+++ app/src/main/java/tmg/hourglass/presentation/home/HomeScreen.kt
@@ -1,7 +1,7 @@
 package tmg.hourglass.presentation.home
 
 import androidx.annotation.StringRes
-import androidx.compose.animation.fadeIn
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -17,9 +17,6 @@ import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Add
-import androidx.compose.material.icons.outlined.Settings
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.runtime.Composable
@@ -28,23 +25,16 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.compose.LifecycleEventEffect
-import tmg.hourglass.BuildConfig
-import java.time.LocalDateTime
-import java.time.ZoneId
-import tmg.hourglass.R
 import tmg.hourglass.domain.enums.CountdownColors
-import tmg.hourglass.domain.enums.CountdownInterpolator
 import tmg.hourglass.domain.enums.CountdownType
 import tmg.hourglass.domain.model.Countdown
 import tmg.hourglass.domain.model.Countdown.Companion.YYYY_MM_DD_FORMAT
+import tmg.hourglass.domain.model.Tag
+import tmg.hourglass.domain.model.TagOrdering
 import tmg.hourglass.presentation.AppTheme
 import tmg.hourglass.presentation.AppThemePreview
 import tmg.hourglass.presentation.PreviewTablet
@@ -52,39 +42,36 @@ import tmg.hourglass.presentation.PreviewTheme
 import tmg.hourglass.presentation.buttons.FloatingActionButton
 import tmg.hourglass.presentation.home.components.Countdown
 import tmg.hourglass.presentation.home.components.Empty
-import tmg.hourglass.presentation.home.components.SortDialog
-import tmg.hourglass.presentation.home.components.label
-import tmg.hourglass.presentation.layouts.MasterDetailsPane
-import tmg.hourglass.presentation.layouts.TitleBar
-import tmg.hourglass.presentation.modify.ModifyScreen
-import tmg.hourglass.presentation.textviews.TextBody2
+import tmg.hourglass.presentation.home.components.Header
+import tmg.hourglass.presentation.home.components.TagHeader
 import tmg.hourglass.presentation.textviews.TextHeader2
+import tmg.hourglass.strings.R
 import tmg.hourglass.strings.R.string
-import java.time.format.DateTimeFormatter
+import java.time.LocalDateTime
+import java.time.ZoneId
 
 @Composable
 internal fun HomeScreenVM(
     paddingValues: PaddingValues,
     windowSize: WindowSizeClass,
+    navigateToAdd: () -> Unit,
+    navigateToModify: (id: String) -> Unit,
+    navigateToSettings: () -> Unit,
     viewModel: HomeViewModel = hiltViewModel()
 ) {
     val uiState = viewModel.uiState.collectAsState()
-
-    LifecycleEventEffect(Lifecycle.Event.ON_RESUME) {
-        viewModel.refresh()
-    }
 
     HomeScreen(
         paddingValues = paddingValues,
         windowSizeClass = windowSize,
         uiState = uiState.value,
-        sortUpdated = viewModel::updateSortOrder,
-        edit = viewModel::edit,
+        tagSortUpdated = viewModel::tagSortUpdated,
+        tagExpanded = viewModel::tagExpanded,
+        untaggedSort = viewModel::untaggedSort,
+        edit = { navigateToModify(it.id) },
         delete = viewModel::delete,
-        openCreateNew = viewModel::createNew,
-        closeDetailPane = viewModel::closeAction,
-        navigateToSettings = viewModel::navigateToSettings,
-        refresh = viewModel::refresh
+        openCreateNew = { navigateToAdd() },
+        navigateToSettings = { navigateToSettings() },
     )
 }
 
@@ -93,103 +80,69 @@ internal fun HomeScreen(
     paddingValues: PaddingValues,
     windowSizeClass: WindowSizeClass,
     uiState: UiState,
-    sortUpdated: (SortOrder) -> Unit,
+    tagSortUpdated: (Tag, TagOrdering) -> Unit,
+    untaggedSort: (TagOrdering) -> Unit,
+    tagExpanded: (Tag, Boolean) -> Unit,
     edit: (Countdown) -> Unit,
     delete: (Countdown) -> Unit,
     openCreateNew: () -> Unit,
-    closeDetailPane: () -> Unit,
     navigateToSettings: () -> Unit,
-    refresh: () -> Unit
 ) {
-    MasterDetailsPane(
+    ListScreen(
+        paddingValues = paddingValues,
+        uiState = uiState,
         windowSizeClass = windowSizeClass,
-        master = {
-            ListScreen(
-                uiState = uiState,
-                windowSizeClass = windowSizeClass,
-                editItem = edit,
-                deleteItem = delete,
-                sortUpdated = sortUpdated,
-                navigateToSettings = navigateToSettings
-            )
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .navigationBarsPadding()
-                    .padding(AppTheme.dimensions.paddingMedium),
-                contentAlignment = Alignment.BottomEnd
-            ) {
-                FloatingActionButton(
-                    label = stringResource(id = string.dashboard_fab_new),
-                    icon = Icons.Outlined.Add,
-                    onClick = openCreateNew
-                )
-            }
-        },
-        detailsShow = uiState.action != null,
-        details = {
-            ModifyScreen(
-                paddingValues = paddingValues,
-                windowSizeClass = windowSizeClass,
-                actionUpClicked = {
-                    closeDetailPane()
-                    refresh()
-                },
-                countdown = (uiState.action as? HomeAction.Modify)?.countdown
-            )
-        },
-        detailsActionUpClicked = closeDetailPane
+        editItem = edit,
+        deleteItem = delete,
+        untaggedSort = untaggedSort,
+        tagSortUpdated = tagSortUpdated,
+        tagExpanded = tagExpanded,
+        navigateToSettings = navigateToSettings
     )
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .navigationBarsPadding()
+            .padding(AppTheme.dimensions.paddingMedium),
+        contentAlignment = Alignment.BottomEnd
+    ) {
+        FloatingActionButton(
+            label = stringResource(id = string.dashboard_fab_new),
+            icon = Icons.Outlined.Add,
+            onClick = openCreateNew
+        )
+    }
 }
 
 @Composable
 internal fun ListScreen(
+    paddingValues: PaddingValues,
     windowSizeClass: WindowSizeClass,
     navigateToSettings: () -> Unit,
-    sortUpdated: (SortOrder) -> Unit,
+    tagSortUpdated: (Tag, TagOrdering) -> Unit,
+    untaggedSort: (TagOrdering) -> Unit,
+    tagExpanded: (Tag, Boolean) -> Unit,
     uiState: UiState,
     editItem: (Countdown) -> Unit,
     deleteItem: (Countdown) -> Unit
 ) {
-    val sortOrderExpanded = remember { mutableStateOf(false) }
-
     LazyVerticalGrid(
         modifier = Modifier.fillMaxSize(),
         columns = GridCells.Adaptive(minSize = 300.dp),
+        contentPadding = PaddingValues(
+            horizontal = AppTheme.dimensions.paddingMedium
+        ),
+        verticalArrangement = Arrangement.spacedBy(AppTheme.dimensions.paddingXSmall),
+        horizontalArrangement = Arrangement.spacedBy(AppTheme.dimensions.paddingXSmall),
         content = {
             item(key = "edgetoedge-header", span = { GridItemSpan(maxLineSpan) }) {
                 Spacer(Modifier.statusBarsPadding())
             }
             item("header", span = { GridItemSpan(maxLineSpan) }) {
-                TitleBar(
+                Header(
                     modifier = Modifier.animateItem(),
-                    title = stringResource(id = R.string.app_name),
-                    overflowActions = {
-                        IconButton(
-                            onClick = {
-                                sortOrderExpanded.value = true
-                            },
-                            content = {
-                                Icon(
-                                    painter = painterResource(R.drawable.ic_sort),
-                                    contentDescription = stringResource(string.menu_sort),
-                                    tint = AppTheme.colors.textPrimary
-                                )
-                            },
-                        )
-                        if (windowSizeClass.widthSizeClass == WindowWidthSizeClass.Compact) {
-                            IconButton(
-                                onClick = navigateToSettings,
-                                content = {
-                                    Icon(
-                                        imageVector = Icons.Outlined.Settings,
-                                        contentDescription = stringResource(string.menu_settings),
-                                        tint = AppTheme.colors.textPrimary
-                                    )
-                                }
-                            )
-                        }
-                    }
+                    windowSizeClass = windowSizeClass,
+                    navigateToSettings = navigateToSettings
                 )
             }
             if (uiState.isEmpty) {
@@ -201,29 +154,56 @@ internal fun ListScreen(
                         Empty()
                     }
                 }
-            } else {
-                item("disclaimer", span = { GridItemSpan(maxLineSpan) }) {
-                    val label = stringResource(uiState.sortOrder.label())
-                    TextBody2(
-                        fontStyle = FontStyle.Italic,
-                        text = stringResource(string.menu_sort_order, label),
-                        modifier = Modifier
-                            .animateItem()
-                            .padding(
-                                horizontal = AppTheme.dimensions.paddingMedium,
-                                vertical = AppTheme.dimensions.paddingXSmall
+            }
+            items(
+                items = uiState.items,
+                key = { it.id },
+                span = {
+                    when (it) {
+                        is ListItem.CountdownItem -> GridItemSpan(1)
+                        is ListItem.TagHeader -> GridItemSpan(maxLineSpan)
+                        is ListItem.UntaggedHeader -> GridItemSpan(maxLineSpan)
+                    }
+                },
+                itemContent = {
+                    when (it) {
+                        is ListItem.CountdownItem -> {
+                            Countdown(
+                                modifier = Modifier.animateItem(),
+                                countdown = it.countdown,
+                                editClicked = editItem,
+                                deleteClicked = deleteItem
                             )
-                    )
+                        }
+                        is ListItem.TagHeader -> {
+                            TagHeader(
+                                name = it.tag.name,
+                                sortClicked = { sort ->
+                                    tagSortUpdated(it.tag, sort)
+                                },
+                                expandClicked = { expanded ->
+                                    tagExpanded(it.tag, expanded)
+                                },
+                                expanded = it.expand ?: false,
+                                showCollapse = true,
+                                modifier = Modifier.padding(top = AppTheme.dimensions.paddingMedium)
+                            )
+                        }
+                        is ListItem.UntaggedHeader -> {
+                            TagHeader(
+                                name = stringResource(R.string.title_all),
+                                sortClicked = { sort ->
+                                    untaggedSort(sort)
+                                },
+                                expandClicked = { },
+                                expanded = false,
+                                showCollapse = false,
+                                modifier = Modifier.padding(top = AppTheme.dimensions.paddingMedium)
+                            )
+                        }
+                    }
                 }
-            }
-            items(uiState.itemsOrdered, key = { it.id }) {
-                Countdown(
-                    modifier = Modifier.animateItem(),
-                    countdown = it,
-                    editClicked = editItem,
-                    deleteClicked = deleteItem
-                )
-            }
+            )
             item("spacer", span = { GridItemSpan(maxLineSpan) }) {
                 Spacer(modifier = Modifier.height(64.dp))
             }
@@ -232,13 +212,6 @@ internal fun ListScreen(
             }
         }
     )
-
-    if (sortOrderExpanded.value) {
-        SortDialog(
-            sortUpdated = sortUpdated,
-            dismissed = { sortOrderExpanded.value = false },
-        )
-    }
 }
 
 @Composable
@@ -262,51 +235,53 @@ private fun PreviewEmpty() {
             paddingValues = PaddingValues.Absolute(),
             windowSizeClass = compactWindowSizeClass,
             uiState = UiState.empty(),
-            sortUpdated = { },
             edit = { },
             delete = { },
             openCreateNew = { },
-            closeDetailPane = { },
-            refresh = { },
+            tagSortUpdated = { _, _ -> },
+            tagExpanded = { _, _ -> },
+            untaggedSort = { },
             navigateToSettings = { }
         )
     }
 }
 
 @PreviewTheme
+@PreviewTablet
 @Composable
-private fun PreviewPopulated() {
+private fun PreviewUntagged() {
     AppThemePreview {
         HomeScreen(
             paddingValues = PaddingValues.Absolute(),
             windowSizeClass = compactWindowSizeClass,
-            uiState = UiState.upcoming(),
-            sortUpdated = { },
+            uiState = UiState.untagged(),
             edit = { },
             delete = { },
             openCreateNew = { },
-            closeDetailPane = { },
-            refresh = { },
+            tagSortUpdated = { _, _ -> },
+            tagExpanded = { _, _ -> },
+            untaggedSort = { },
             navigateToSettings = { }
         )
     }
 }
 
 
+@PreviewTheme
 @PreviewTablet
 @Composable
-private fun PreviewPopulatedExpanded() {
+private fun PreviewTagged() {
     AppThemePreview {
         HomeScreen(
             paddingValues = PaddingValues.Absolute(),
             windowSizeClass = expandedWindowSizeClass,
-            uiState = UiState.upcoming(),
-            sortUpdated = { },
+            uiState = UiState.tagged(),
             edit = { },
             delete = { },
             openCreateNew = { },
-            closeDetailPane = { },
-            refresh = { },
+            tagSortUpdated = { _, _ -> },
+            tagExpanded = { _, _ -> },
+            untaggedSort = { },
             navigateToSettings = { }
         )
     }
@@ -318,22 +293,43 @@ private val compactWindowSizeClass: WindowSizeClass =
 private val expandedWindowSizeClass: WindowSizeClass =
     WindowSizeClass.calculateFromSize(DpSize(910.dp, 600.dp))
 
-private fun UiState.Companion.empty(
-    withAction: Boolean = false
-): UiState = UiState(
-    items = emptyList(),
-    sortOrder = SortOrder.ALPHABETICAL,
-    action = if (withAction) HomeAction.Add else null
+private fun UiState.Companion.empty(): UiState = UiState(
+    items = emptyList()
 )
 
-private fun UiState.Companion.upcoming(
-    withAction: Boolean = false,
-    withUpcoming: Boolean = true,
-): UiState = UiState(
-    items = if (withUpcoming) listOf(fakeCountdownUpcoming, fakeCountdownUpcoming.copy(id = "upcoming_2")) else emptyList(),
-    sortOrder = SortOrder.ALPHABETICAL,
-    action = if (withAction) HomeAction.Add else null
-)
+private fun UiState.Companion.untagged(): UiState = UiState(
+    items = listOf(
+        ListItem.CountdownItem(
+            countdown = fakeCountdownUpcoming
+        ),
+        ListItem.CountdownItem(
+            countdown = fakeCountdownUpcoming.copy(id = "upcoming_2")
+        )
+    ))
+
+private fun UiState.Companion.tagged(): UiState = UiState(
+    items = listOf(
+        ListItem.TagHeader(
+            tag = Tag("tag1", "A", "", TagOrdering.ALPHABETICAL),
+            expand = true
+        ),
+        ListItem.CountdownItem(
+            countdown = fakeCountdownUpcoming.copy(id = "upcoming_1")
+        ),
+        ListItem.CountdownItem(
+            countdown = fakeCountdownUpcoming.copy(id = "upcoming_2")
+        ),
+        ListItem.TagHeader(
+            tag = Tag("tag2", "A", "", TagOrdering.ALPHABETICAL),
+            expand = true
+        ),
+        ListItem.CountdownItem(
+            countdown = fakeCountdownUpcoming.copy(id = "upcoming_3")
+        ),
+        ListItem.CountdownItem(
+            countdown = fakeCountdownUpcoming.copy(id = "upcoming_4")
+        )
+    ))
 
 private val fakeCountdownUpcoming = Countdown.Static(
     id = "upcoming_1",
@@ -344,5 +340,6 @@ private val fakeCountdownUpcoming = Countdown.Static(
     end = LocalDateTime.now(ZoneId.of("UTC")).plusDays(2).format(YYYY_MM_DD_FORMAT),
     startValue = "0",
     endValue = "10000",
-    countdownType = CountdownType.DAYS
+    countdownType = CountdownType.DAYS,
+    tag = null
 )
